pyproject.toml
[project]
name = "aia-async"
version = "0.1.0"
description = ""
authors = [
    {name = "Bogdan Opanchuk", email = "bogdan@opanchuk.net"},
]
dependencies = [
    "anyio>=3",
    "cryptography>=36",
    "pyopenssl>=22"
]
requires-python = ">=3.8"
license-expression = "MIT"

[project.urls]
homepage = ""

[tool]
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
